# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "daw13.2022@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/peticiones": {
            "get": {
                "tags": [
                    "getPeticiones"
                ],
                "summary": "Get list of peticiones",
                "description": "Returns list of peticiones",
                "operationId": "getPeticiones",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/mispeticiones": {
            "get": {
                "tags": [
                    "getMispeticiones"
                ],
                "summary": "Peticiones que ha realizado un usuario",
                "description": "Peticiones que ha realizado un usuario",
                "operationId": "getMispeticiones",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/peticiones/{peticiones}": {
            "get": {
                "tags": [
                    "getShow"
                ],
                "summary": "Mostrar una peticione determinada",
                "description": "Mostrar una peticione determinada",
                "operationId": "getShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Peticiones id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:peticiones",
                            "read:peticiones"
                        ]
                    }
                ]
            }
        }
    }
}